Revert to conditional patchelfing

+++ default.nix
@@ -67,12 +67,11 @@ let
               for binary in dfx ic-ref ic-starter mo-doc mo-ide moc replica; do
                 ${self.lib.optionalString self.stdenv.isLinux ''
                 local BINARY="$out/cache/''${binary}"
-                # test -f "$BINARY" || continue
-                # local IS_STATIC=$(ldd "''${BINARY}" | grep 'not a dynamic executable')
-                # local USE_LIB64=$(ldd "''${BINARY}" | grep '/lib64/ld-linux-x86-64.so.2')
+                test -f "$BINARY" || continue
+                local IS_STATIC=$(ldd "''${BINARY}" | grep 'not a dynamic executable')
+                local USE_LIB64=$(ldd "''${BINARY}" | grep '/lib64/ld-linux-x86-64.so.2')
                 chmod +rw "''${BINARY}"
-                # test -n "$IS_STATIC" || test -z "$USE_LIB64" || patchelf --set-interpreter "''${LD_LINUX_SO}" "''${BINARY}"
-                patchelf --set-interpreter "''${LD_LINUX_SO}" "''${BINARY}"
+                test -n "$IS_STATIC" || test -z "$USE_LIB64" || patchelf --set-interpreter "''${LD_LINUX_SO}" "''${BINARY}"
                 ''}
                 ln -s $out/cache/$binary $out/bin/$binary
               done
